+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## New Features
 
-None...
+1. Added Missile PowerUp.
+2. When you shot a missile, a small red bullet will be shot out. When this bullet hits a wall, it will explode and destroy all rockets within a fixed radius from it regardless of the owner of the missile.
 
 ## Changes
 
-1. Fixed bug where Player 2 would move faster than Player 1.
+None...